refactor: Address review comments for ivyless publishLocal
